HelloWorldWideGUI Program (Part 3)
Learning to add interactive elements to the stage to be presented in the GUI window to make things more intricate and fuller variety for the user, specifically with through the example of buttons. This will be demonstrated through creating a standalone GUI program that will showcase a list of different languages as buttons. Clicking on a button will show a line of text on how to say, "Hello World!" in its respective language, and an exit button is located at the button of the list that closes the application if the user wishes to. As well as a few other elements to make the program complete.